backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
numpy
pillow
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
google-generativeai
